+++ test/LionWeb.Core.Test/NodeApi/Dynamic/StructuredDataTypesTests.cs
@@ -173,7 +173,8 @@ public class StructuredDataTypesTests
             .OfType(sdt)
             .IsOptional(true);
 
-        var chunk = new Serializer(lionWebVersion).SerializeToChunk([language]);
+        var chunk = new SerializerBuilder().WithLionWebVersion(lionWebVersion).Build()
+            .SerializeToChunk([language]);
         Assert.ThrowsException<UnsupportedStructuredDataTypeException>(() =>
             new LanguageDeserializer(lionWebVersion).Deserialize(chunk));
     }